+++ template-sample-app/template/src/components/views/stores/StoresDetail.js
@@ -14,10 +14,10 @@ import {
 import NavigateNextIcon from '@material-ui/icons/NavigateNext';
 
 import { WrapperWidgetUI, FormulaWidgetUI, HistogramWidgetUI } from '@carto/react/ui';
-import { selectSourceById, setViewState } from '@carto/react/redux';
+import { updateLayer, selectSourceById, setViewState } from '@carto/react/redux';
 
 import { getStore, getRevenuePerMonth } from 'models/StoreModel';
-import { MONTHS_LABELS } from './constants';
+import { LAYER_ID, MONTHS_LABELS } from './constants';
 import { Isochrone } from 'components/common/Isochrone';
 import { currencyFormatter } from 'utils/formatter';
 import { setBottomSheetOpen, setError } from 'config/appSlice';
@@ -71,9 +71,13 @@ export default function StoresDetail() {
         dispatch(setError(`getRevenuePerMonth error: ${error.message}`));
       });
 
+    // Set selected store on the layer
+    dispatch(updateLayer(LAYER_ID, { selectedStore: id }));
+    
     dispatch(setBottomSheetOpen(true));
 
-    return function cleanup() {
+    return () => {
+      dispatch(updateLayer(LAYER_ID, { selectedStore: null }));
       abortController.abort();
     };
   }, [dispatch, source, id, location.state]);